Switch backend to Postgres; add docker-compose with bundled db
Add psycopg[binary] so SQLAlchemy can talk to Postgres, and a
docker-compose stack with a Postgres `db` service the app targets by
default. SQLite remains available (override DATABASE_URL) and is still
used for the local test suite.

Fix a Postgres-only overflow: scores.created_at / active_games.updated_at
store millisecond epochs (~1.78e12), which exceed 32-bit INTEGER. Use
BigInteger for those columns. SQLite's flexible INTEGER had masked this.

Consolidate env config to the repo-root .env (gitignored) + .env.example;
the app service reads it via env_file. Remove backend/.env.example.

+++ backend/app/db.py
@@ -10,7 +10,7 @@ engine is created only when :func:`make_engine` is called (i.e. when a
 ``DATABASE_URL`` is configured).
 """
 
-from sqlalchemy import JSON, Boolean, Integer, String, create_engine
+from sqlalchemy import JSON, BigInteger, Boolean, Integer, String, create_engine
 from sqlalchemy.engine import Engine
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 from sqlalchemy.pool import StaticPool
@@ -45,7 +45,9 @@ class ScoreTable(Base):
     username: Mapped[str] = mapped_column(String, nullable=False)
     mode: Mapped[str] = mapped_column(String, nullable=False, index=True)
     score: Mapped[int] = mapped_column(Integer, nullable=False)
-    created_at: Mapped[int] = mapped_column(Integer, nullable=False)
+    # Millisecond Unix epoch — exceeds 32-bit INTEGER range, so use BigInteger
+    # (Postgres rejects it as int4; SQLite's flexible INTEGER masked this).
+    created_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
 
 
 class ActiveGameTable(Base):
@@ -57,7 +59,8 @@ class ActiveGameTable(Base):
     mode: Mapped[str] = mapped_column(String, nullable=False)
     score: Mapped[int] = mapped_column(Integer, nullable=False)
     state: Mapped[dict] = mapped_column(JSON, nullable=False)
-    updated_at: Mapped[int] = mapped_column(Integer, nullable=False)
+    # Millisecond Unix epoch — see ScoreTable.created_at; needs BigInteger.
+    updated_at: Mapped[int] = mapped_column(BigInteger, nullable=False)
     # Pinned games are exempt from stale-pruning (used for demo/seed games).
     pinned: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
